requirements.txt
playwright
python-dotenv
apscheduler
